==> RW/Scripts/1. RW_airports_merging.ipynb <==
import os
import sys
import numpy as np
import pandas as pd
from geopy import Point
from geopy import distance
pd.set_option('display.max_columns', 999)

# load datasets
rw = pd.read_csv("..//Datasets/NEW/RW_flights.csv", parse_dates=['date'])
rw_info = pd.read_csv("..//Datasets/NEW/RW_services_info.csv")
airports = pd.read_csv("..//Datasets/Airports/BTS_Airports_LAT_LON.csv")
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# number of existing tail numbers with flight history
rw.tail_number.nunique() 

# number of unique air medical services
rw.Owner.nunique()

# The percentage of the total tail numbers
print(round(rw.tail_number.nunique() / rw_info.Tail_number.nunique() * 100, 2), "%")

# count of training flights 
len(rw.loc[rw.origin == rw.destination])

rw.info()

rw.isna().sum()

#### Data preprocessing
---

# check duplicates airport_codes
airport_codes.duplicated(subset=['gps_code']).sum()

# check duplicates airport_codes
airport_codes.duplicated(subset=['local_code']).sum()

# Remove useless words from rw
rw = rw.replace(regex={'Near ':'', 'First seen ':'', 'Last seen ':'', 'result unknown': np.nan, 'En Route':None, 'unknown': np.nan,
                       'n/a': np.nan, 'Unknown':None, 'Diverted':np.nan, '¬†':' ', 'Scheduled': None, 'Cancelled': np.nan})

# count of incomplete flights 
len(rw.loc[(rw.arrival.isna()) | (rw.duration.isna())])

# split origin_Latitude and origin_Longitude
rw['origin_Latitude'] = rw[rw.origin_location.str.startswith('L ')].origin_location.replace(regex={'L ':''}).str.split(" ", n = 1, expand = True)[0]
rw['origin_Longitude'] = rw[rw.origin_location.str.startswith('L ')].origin_location.replace(regex={'L ':''}).str.split(" ", n = 1, expand = True)[1]

# split destination_Latitude and destination_Longitude
rw['destination_Latitude'] = rw[rw.destination_location.str.startswith('L ', na=False)].destination_location.replace(regex={'L ':''}).str.split(" ", n = 1, expand = True)[0]
rw['destination_Longitude'] = rw[rw.destination_location.str.startswith('L ', na=False)].destination_location.replace(regex={'L ':''}).str.split(" ", n = 1, expand = True)[1]

# convert to numeric
rw['origin_Latitude'] = pd.to_numeric(rw.origin_Latitude, errors='coerce')
rw['origin_Longitude'] = pd.to_numeric(rw.origin_Longitude, errors='coerce')

rw['destination_Latitude'] = pd.to_numeric(rw.destination_Latitude, errors='coerce')
rw['destination_Longitude'] = pd.to_numeric(rw.destination_Longitude, errors='coerce')

# split ICAO_code
rw['ICAO_code_origin'] = rw[rw.origin_Latitude.isna()].origin_location.str.rsplit(" - ", n = 1, expand = True)[1]
rw['ICAO_code_destination'] = rw[rw.destination_Latitude.isna()].destination_location.str.rsplit(" - ", n = 1, expand = True)[1]

# replace dual ICAO code
rw.loc[rw.ICAO_code_origin.str.contains(' / ', na=False), 'ICAO_code_origin'] = rw.loc[rw.ICAO_code_origin.str.contains(' / ', na=False), 'ICAO_code_origin'].str.split(" / ", expand=True, n=1)[0]

rw.loc[rw.ICAO_code_destination.str.contains(' / ', na=False), 'ICAO_code_destination'] = rw.loc[rw.ICAO_code_destination.str.contains(' / ', na=False), 'ICAO_code_destination'].str.split(" / ", expand=True, n=1)[0]

#### Origin Location
---

# merge origin airport information 1 time
rw = rw.merge(airport_codes[['ident']].add_suffix('_origin_1'),
         how='left', left_on='ICAO_code_origin', right_on='ident_origin_1')

# Switch dual ICAO codes
rw.loc[(rw.ICAO_code_origin.notnull()) & (rw.ident_origin_1.isna()) & rw.origin.str.contains(' / '), 'ICAO_code_origin'] = rw.loc[(rw.ICAO_code_origin.notnull()) & (rw.ident_origin_1.isna()) & rw.origin.str.contains(' / '), 'origin'].str.split(' / ', n=1, expand=True)[1].str.replace(')', '')

# merge origin airport information 2 time
rw = rw.merge(airport_codes[['ident']].add_suffix('_origin_2'),
         how='left', left_on='ICAO_code_origin', right_on='ident_origin_2')

# merge similar columns
rw.loc[(rw.ICAO_code_origin.notnull()) & (rw.ident_origin_1.isna()) & (rw.origin.str.contains(' / ')), 'ident_origin_1'] = rw.loc[(rw.ICAO_code_origin.notnull()) & (rw.ident_origin_1.isna()) & (rw.origin.str.contains(' / ')), 'ident_origin_2']

# Drop duplicated column
rw = rw.drop(['ident_origin_2'], axis=1)

# load airport_codes again
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# drop duplicates in gps_code
airport_codes = airport_codes.drop_duplicates(subset=['gps_code'])

# merge origin airport information 3 time
rw = rw.merge(airport_codes[['gps_code', 'ident']].add_suffix('_origin_2'),
         how='left', left_on='ICAO_code_origin', right_on='gps_code_origin_2')

# merge similar columns
rw.loc[(rw.ICAO_code_origin.notnull()) & (rw.ident_origin_1.isna()), 'ident_origin_1'] = rw.loc[(rw.ICAO_code_origin.notnull()) & (rw.ident_origin_1.isna()), 'ident_origin_2']

# Drop duplicated columns
rw = rw.drop(['ident_origin_2', 'gps_code_origin_2'], axis=1)

# load airport_codes again
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# drop duplicates in local_code
airport_codes = airport_codes.drop_duplicates(subset=['local_code'])

# merge origin airport information 4 times
rw = rw.merge(airport_codes[['local_code', 'ident']].add_suffix('_origin_2'),
         how='left', left_on='ICAO_code_origin', right_on='local_code_origin_2')

# merge similar columns
rw.loc[(rw.ICAO_code_origin.notnull()) & (rw.ident_origin_1.isna()), 'ident_origin_1'] = rw.loc[(rw.ICAO_code_origin.notnull()) & (rw.ident_origin_1.isna()), 'ident_origin_2']

# Drop duplicated columns
rw = rw.drop(['ident_origin_2', 'local_code_origin_2'], axis=1)

# load airport_codes again
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# split coordinates and convert to numeric
airport_codes["Latitude"] = pd.to_numeric(airport_codes.coordinates.str.split(', ', n=1, expand=True)[1])
airport_codes["Longitude"] = pd.to_numeric(airport_codes.coordinates.str.split(', ', n=1, expand=True)[0])

airport_codes = airport_codes.replace(regex={r' Heliport': '', r' Airpark':'', r' Airport': '', r' Regional':'', r' Municipal':'', r' International':'',
                                            r' LRRS':''})

# add state to the name
airport_codes.loc[:, 'name'] = airport_codes.loc[:, 'name'] + ", " + airport_codes.iso_region.str.split('-', expand=True)[1]

# drop duplicates
airport_codes = airport_codes.drop_duplicates(subset=['name'])

# merge origin airport information 5 times
rw = rw.merge(airport_codes[['name', 'ident']].add_suffix('_origin_2'),
         how='left', left_on='origin', right_on='name_origin_2')

# merge similar columns
rw.loc[(rw.ICAO_code_origin.isna()), 'ident_origin_1'] = rw.loc[(rw.ICAO_code_origin.isna()), 'ident_origin_2'] 
rw.loc[(rw.ICAO_code_origin.isna()), 'ICAO_code_origin'] = rw.loc[(rw.ICAO_code_origin.isna()), 'ident_origin_2'] 

# Drop duplicated columns
rw = rw.drop(['ident_origin_2', 'name_origin_2'], axis=1)

# load airport_codes again
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# split coordinates and convert to numeric
airport_codes["Latitude"] = pd.to_numeric(airport_codes.coordinates.str.split(', ', n=1, expand=True)[1])
airport_codes["Longitude"] = pd.to_numeric(airport_codes.coordinates.str.split(', ', n=1, expand=True)[0])

# remove all extra words
airport_codes = airport_codes.replace(regex={r' Heliport': '', r' Airpark':'', r' Airport': '', r' Regional':'', r' Municipal':'', r' International':'',
                                            r' LRRS':''})
# add state to municipality
airport_codes.loc[:, 'municipality'] = airport_codes.loc[:, 'municipality'] + ", " + airport_codes.iso_region.str.split('-', expand=True)[1]

# drop duplicates
airport_codes = airport_codes.drop_duplicates(subset=['municipality'])

# merge origin airport information 6 times
rw = rw.merge(airport_codes[['municipality', 'ident']].add_suffix('_origin_2'),
         how='left', left_on='origin', right_on='municipality_origin_2')

# merge similar columns
rw.loc[(rw.ICAO_code_origin.isna()), 'ident_origin_1'] = rw.loc[(rw.ICAO_code_origin.isna()), 'ident_origin_2'] 
rw.loc[(rw.ICAO_code_origin.isna()), 'ICAO_code_origin'] = rw.loc[(rw.ICAO_code_origin.isna()), 'ident_origin_2'] 

# Drop duplicated columns
rw = rw.drop(['ident_origin_2', 'municipality_origin_2'], axis=1)

rw.info()

rw.isna().sum()

#### Destination Location 
---

# load airport_codes again
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# merge destination airport information 1 time
rw = rw.merge(airport_codes[['ident']].add_suffix('_destination_1'),
         how='left', left_on='ICAO_code_destination', right_on='ident_destination_1')

# Switch dual ICAO codes
rw.loc[(rw.ICAO_code_destination.notnull()) & (rw.ident_destination_1.isna()) & rw.destination.str.contains(' / '), 'ICAO_code_destination'] = rw.loc[(rw.ICAO_code_destination.notnull()) & (rw.ident_destination_1.isna()) & rw.destination.str.contains(' / '), 'destination'].str.split(' / ', n=1, expand=True)[1].str.replace(')', '')

# merge destination airport information 2 time
rw = rw.merge(airport_codes[['ident']].add_suffix('_destination_2'),
         how='left', left_on='ICAO_code_destination', right_on='ident_destination_2')

# merge similar columns
rw.loc[(rw.ICAO_code_destination.notnull()) & (rw.ident_destination_1.isna()) & (rw.destination.str.contains(' / ')), 'ident_destination_1'] = rw.loc[(rw.ICAO_code_destination.notnull()) & (rw.ident_destination_1.isna()) & (rw.destination.str.contains(' / ')), 'ident_destination_2']

# Drop duplicated column
rw = rw.drop(['ident_destination_2'], axis=1)

# load airport_codes again
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# drop duplicates in gps_code
airport_codes = airport_codes.drop_duplicates(subset=['gps_code'])

# merge destination airport information 3 time
rw = rw.merge(airport_codes[['gps_code', 'ident']].add_suffix('_destination_2'),
         how='left', left_on='ICAO_code_destination', right_on='gps_code_destination_2')

# merge similar columns
rw.loc[(rw.ICAO_code_destination.notnull()) & (rw.ident_destination_1.isna()), 'ident_destination_1'] = rw.loc[(rw.ICAO_code_destination.notnull()) & (rw.ident_destination_1.isna()), 'ident_destination_2']

# Drop duplicated columns
rw = rw.drop(['ident_destination_2', 'gps_code_destination_2'], axis=1)

# load airport_codes again
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# drop duplicates in local_code
airport_codes = airport_codes.drop_duplicates(subset=['local_code'])

# merge destination airport information 4 times
rw = rw.merge(airport_codes[['local_code', 'ident']].add_suffix('_destination_2'),
         how='left', left_on='ICAO_code_destination', right_on='local_code_destination_2')

# merge similar columns
rw.loc[(rw.ICAO_code_destination.notnull()) & (rw.ident_destination_1.isna()), 'ident_destination_1'] = rw.loc[(rw.ICAO_code_destination.notnull()) & (rw.ident_destination_1.isna()), 'ident_destination_2']

# Drop duplicated columns
rw = rw.drop(['ident_destination_2', 'local_code_destination_2'], axis=1)

# load airport_codes again
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# split coordinates and convert to numeric
airport_codes["Latitude"] = pd.to_numeric(airport_codes.coordinates.str.split(', ', n=1, expand=True)[1])
airport_codes["Longitude"] = pd.to_numeric(airport_codes.coordinates.str.split(', ', n=1, expand=True)[0])

airport_codes = airport_codes.replace(regex={r' Heliport': '', r' Airpark':'', r' Airport': '', r' Regional':'', 
                                             r' Municipal':'', r' International':'', r' LRRS':''})

airport_codes.loc[:, 'name'] = airport_codes.loc[:, 'name'] + ", " + airport_codes.iso_region.str.split('-', expand=True)[1]
airport_codes.loc[:, 'municipality'] = airport_codes.loc[:, 'municipality'] + ", " + airport_codes.iso_region.str.split('-', expand=True)[1]

# drop duplicates
airport_codes = airport_codes.drop_duplicates(subset=['municipality'])
airport_codes = airport_codes.drop_duplicates(subset=['name'])

# merge destination airport information 5 times
rw = rw.merge(airport_codes[['name', 'ident']].add_suffix('_destination_2'),
         how='left', left_on='destination', right_on='name_destination_2')

# merge similar columns
rw.loc[(rw.ICAO_code_destination.isna()), 'ident_destination_1'] = rw.loc[(rw.ICAO_code_destination.isna()), 'ident_destination_2'] 
rw.loc[(rw.ICAO_code_destination.isna()), 'ICAO_code_destination'] = rw.loc[(rw.ICAO_code_destination.isna()), 'ident_destination_2'] 

# Drop duplicated columns
rw = rw.drop(['ident_destination_2', 'name_destination_2'], axis=1)

# merge destination airport information 6 times
rw = rw.merge(airport_codes[['municipality', 'ident']].add_suffix('_destination_2'),
         how='left', left_on='destination', right_on='municipality_destination_2')

# merge similar columns
rw.loc[(rw.ICAO_code_destination.isna()), 'ident_destination_1'] = rw.loc[(rw.ICAO_code_destination.isna()), 'ident_destination_2'] 
rw.loc[(rw.ICAO_code_destination.isna()), 'ICAO_code_destination'] = rw.loc[(rw.ICAO_code_destination.isna()), 'ident_destination_2'] 

# Drop duplicated columns
rw = rw.drop(['ident_destination_2', 'municipality_destination_2'], axis=1)

rw.info()

rw.isna().sum()

#### Coordinates merging
---

# load airport_codes again
airport_codes = pd.read_csv("..//Datasets/Airports/airport-codes.csv")

# split coordinates and convert to numeric
airport_codes["Latitude"] = pd.to_numeric(airport_codes.coordinates.str.split(', ', n=1, expand=True)[1])
airport_codes["Longitude"] = pd.to_numeric(airport_codes.coordinates.str.split(', ', n=1, expand=True)[0])

# merge by origin
rw = rw.merge(airport_codes.add_suffix('_origin'),
         how='left', left_on='ident_origin_1', right_on='ident_origin')

# replace origin_Latitude and origin_Longitude
rw.loc[rw.origin_Latitude.isna(), 'origin_Latitude'] = rw.loc[rw.origin_Latitude.isna(), 'Latitude_origin']
rw.loc[rw.origin_Longitude.isna(), 'origin_Longitude'] = rw.loc[rw.origin_Longitude.isna(), 'Longitude_origin']

# drop all useless columns
rw = rw.drop(['coordinates_origin', 'Latitude_origin', 'Longitude_origin', 'continent_origin', 'elevation_ft_origin', 'type_origin',
         'local_code_origin', 'gps_code_origin', 'municipality_origin', 'iso_region_origin', 'ident_origin', 'iata_code_origin'], axis=1)

# merge by destination
rw = rw.merge(airport_codes.add_suffix('_destination'),
         how='left', left_on='ident_destination_1', right_on='ident_destination')

# replace destination_Latitude and destination_Longitude
rw.loc[rw.destination_Latitude.isna(), 'destination_Latitude'] = rw.loc[rw.destination_Latitude.isna(), 'Latitude_destination']
rw.loc[rw.destination_Longitude.isna(), 'destination_Longitude'] = rw.loc[rw.destination_Longitude.isna(), 'Longitude_destination']

# drop all useless columns
rw = rw.drop(['coordinates_destination', 'Latitude_destination', 'Longitude_destination', 'continent_destination', 'elevation_ft_destination', 'type_destination',
         'local_code_destination', 'gps_code_destination', 'municipality_destination', 'iso_region_destination', 'ident_destination', 'iata_code_destination'], axis=1)

rw.isna().sum()

rw.info()

#### Additional Information
---

# add Owner columns
rw = rw.merge(rw_info[['Tail_number','Owner']], how='left', left_on='tail_number', right_on='Tail_number').drop('Tail_number', axis=1)

# calculate the distance
for i in range(len(rw)):
    try:
        p1 = Point(str(rw.origin_Latitude[i]) + ' ' + str(rw.origin_Longitude[i]))
        p2 = Point(str(rw.destination_Latitude[i]) + ' ' + str(rw.destination_Longitude[i]))
    
        rw.loc[i, 'distance_mi'] = distance.distance(p1,p2).miles
    except:
        pass

# convert duration to minutes
rw.duration = pd.to_numeric(rw.duration.str.split(":", expand=True)[0]) * 60 + pd.to_numeric(rw.duration.str.split(":", expand=True)[1])

# split states 
rw['origin_state'] = rw.origin.str.split(', ', n=1, expand=True)[1]
rw.loc[rw.origin_state.isna(), 'origin_state'] = rw.origin_location.str.split(', ', n=1, expand=True)[1].str.rsplit(') ', n=1, expand=True)[0]
rw.loc[rw.origin_state.str.contains(' ', na=False), 'origin_state'] = rw.origin_location.str.rsplit(', ', n=1, expand=True)[1].str.split(')', n=1, expand=True)[0]

rw['destination_state'] = rw.destination.str.split(', ', n=1, expand=True)[1]
rw.loc[rw.destination_state.isna(), 'destination_state'] = rw.destination_location.str.split(', ', n=1, expand=True)[1].str.rsplit(') ', n=1, expand=True)[0]
rw.loc[rw.destination_state.str.contains(' ', na=False), 'destination_state'] = rw.destination_location.str.rsplit(', ', n=1, expand=True)[1].str.split(')', n=1, expand=True)[0]

# split cities
rw['city_origin'] = rw.origin_location.str.rsplit(" (", n=1, expand=True)[1].str.rsplit(", ", n=1, expand=True)[0]
rw.loc[rw.city_origin.isna(), 'city_origin'] = rw.loc[rw.origin.str.contains(', ', na=False), 'origin'].str.split(', ', expand=True)[0]
rw.loc[rw.city_origin.str.contains(' - ', na=False), 'city_origin'] = None

rw['city_destination'] = rw.destination_location.str.rsplit(" (", n=1, expand=True)[1].str.rsplit(", ", n=1, expand=True)[0]
rw.loc[rw.city_destination.isna(), 'city_destination'] = rw.loc[rw.destination.str.contains(', ', na=False), 'destination'].str.split(', ', expand=True)[0]
rw.loc[rw.city_destination.str.contains(' - ', na=False), 'city_origin'] = None

# add shift
rw['departure_shift'] = np.where((pd.to_datetime(rw.departure.str.split('M', n=1, expand=True)[0] + 'M', errors='coerce').dt.time.astype('str') < '19:00:00') &\
                       (pd.to_datetime(rw.departure.str.split('M', n=1, expand=True)[0] + 'M', errors='coerce').dt.time.astype('str') > '07:00:00'), 'day', 'night')

rw['arrival_shift'] = np.where((pd.to_datetime(rw.arrival.str.split('M', n=1, expand=True)[0] + 'M', errors='coerce').dt.time.astype('str') < '19:00:00') &\
                       (pd.to_datetime(rw.arrival.str.split('M', n=1, expand=True)[0] + 'M', errors='coerce').dt.time.astype('str') > '07:00:00'), 'day', 'night')

# rename columns and drop unnecessary columns
rw = rw.rename({'ident_origin_1': 'origin_code', 'ident_destination_1': 'destination_code', 
           'name_origin':'airport_origin', 'iso_country_origin': 'country_origin',
          'name_destination':'airport_destination', 'iso_country_destination': 'country_destination'}, axis=1)\
       .drop(['origin_location', 'destination_location', 'ICAO_code_origin', 'ICAO_code_destination'], axis=1)

rw.info()

# rearrange columns
rw = rw.iloc[:,:3].join([rw.iloc[:,12], rw.iloc[:,3], rw.iloc[:,13], rw.iloc[:,4], rw.iloc[:,5:8], rw.iloc[:,19], rw.iloc[:,18], 
                    rw.iloc[:,8:12], rw.iloc[:,14], rw.iloc[:,22], rw.iloc[:,20], rw.iloc[:,15],
                    rw.iloc[:,16], rw.iloc[:,23], rw.iloc[:,21], rw.iloc[:,17], rw.iloc[:,24:26]])

rw.head()

rw.isna().sum()

# drop all imcomplete records
rw = rw.loc[(rw.arrival.notnull()) & (rw.duration.notnull()) & (rw.country_origin == 'US') & (rw.country_destination == 'US')\
            & (rw.duration != 0) & (rw.origin != rw.destination) & (rw.distance_mi != 0)]

rw.info()

# import to csv
rw.to_csv("..//Datasets/NEW/RW_with_airports.csv", index=False)
