==> RW/Scripts/2. RW METAR data merging.ipynb <==
import os
import sys
import pandas as pd
import numpy as np
sys.path.append(os.getcwd() + '/Modules/')
from utc_converter import utc_converter
from metar_decoder import metar_decoder
pd.set_option('display.max_columns', None)

#### Data preprocessing
---

# load data frame
rw = pd.read_csv("..//Datasets/NEW/RW_with_airports.csv")

# add time with UTC format 
rw = utc_converter(rw)

rw.head()

# convert date to datetime
rw.date = pd.to_datetime(rw.date, format='%m/%d/%y')

# split ICAO_code
rw['ICAO_origin'] = rw.origin.str.rsplit("(", n = 1, expand = True)[1].str.split(")", n = 1, expand = True)[0]
rw['ICAO_destination'] = rw.destination.str.rsplit("(", n = 1, expand = True)[1].str.split(")", n = 1, expand = True)[0]

# Split dual ICAO and take the last one
rw.loc[rw.ICAO_origin.str.contains(' / ', na=False), 'ICAO_origin'] = rw.loc[rw.ICAO_origin.str.contains(' / ', na=False), 'ICAO_origin']\
                                                                        .str.split(" / ", expand=True, n=1)[1]

rw.loc[rw.ICAO_destination.str.contains(' / ', na=False), 'ICAO_destination'] = rw.loc[rw.ICAO_destination.str.contains(' / ', na=False), 'ICAO_destination']\
                                                                                  .str.split(" / ", expand=True, n=1)[1]

# replace null ICAO code by airport code
rw.loc[rw.ICAO_origin.isna(), 'ICAO_origin'] = rw.loc[rw.ICAO_origin.isna(), 'origin_code'] 
rw.loc[rw.ICAO_destination.isna(), 'ICAO_destination'] = rw.loc[rw.ICAO_destination.isna(), 'destination_code']

# Transform aita code to icao
rw.loc[~rw.ICAO_origin.str.startswith('K', na=False), 'ICAO_origin'] = 'K' + rw.loc[~rw.ICAO_origin.str.startswith('K', na=False), 'ICAO_origin']
rw.loc[~rw.ICAO_destination.str.startswith('K', na=False), 'ICAO_destination'] = 'K' + rw.loc[~rw.ICAO_destination.str.startswith('K', na=False), 'ICAO_destination']

# extract hour from dep_UTC_time and assign a new variable
rw['dep_UTC_hour'] = pd.to_datetime(rw.dep_UTC_time.astype('str'), format='%H:%M:%S', errors='coerce').dt.hour.astype('str')
rw['arr_UTC_hour'] = pd.to_datetime(rw.arr_UTC_time.astype('str'), format='%H:%M:%S', errors='coerce').dt.hour.astype('str')

# sort by 'tail_number', 'date', 'dep_UTC_hour'
rw = rw.sort_values(['tail_number', 'date', 'dep_UTC_hour']).reset_index(drop=True)

# inspect data
rw

#### METAR data merging for Origin Location
---

# # scrape METAR data for Origin
# from metar_scraper_origin import metar_scraper_origin
# metar_scraper_origin(rw)

# load data frame
metar_origin = pd.read_csv("..//Datasets/NEW/METAR/RW_METAR_origin.csv")

# split date and time. Extract only hours  
metar_origin['hour'] = pd.to_datetime(metar_origin.date, format='%d/%m/%Y %H:%M').dt.hour.astype('str')
metar_origin['date'] = pd.to_datetime(metar_origin.date, format='%d/%m/%Y %H:%M').dt.date.astype('str')

# convert date to string
rw.date = rw.date.astype('str')

# sort by 'tail_number', 'date', 'hour' and drop duplicates
metar_origin = metar_origin.sort_values(['tail_number', 'date', 'hour']).drop_duplicates(subset=['tail_number', 'ICAO_origin', 'date', 'hour'])\
                           .reset_index(drop=True)

# inspect data
metar_origin.head()

# merge by 'tail_number', 'date', 'ICAO_origin', 'dep_UTC_hour'
rw = rw.merge(metar_origin.add_suffix('_origin'), how='left', left_on=['tail_number', 'date', 'ICAO_origin', 'dep_UTC_hour'],
        right_on=['tail_number_origin', 'date_origin', 'ICAO_origin_origin', 'hour_origin'])\
        .drop(['ICAO_origin', 'dep_UTC_hour','tail_number_origin', 'ICAO_origin_origin', 'date_origin', 'hour_origin'], axis=1)

# check for null values
rw.METAR_origin.isna().sum()

# percent of merged metar origin data
print(round(len(rw.loc[rw.METAR_origin.notnull()])/ len(rw) * 100,2), "%")

rw.loc[rw.METAR_origin.isna()]

#### METAR data merging for Destination Location
---

# # scrape METAR data for Origin
# from metar_scraper_destination import metar_scraper_destination
# metar_scraper_destination(rw)

# load data frame
metar_destination = pd.read_csv("../Datasets/NEW/METAR/RW_METAR_destination.csv")

# split date and time. Extract only hours  
metar_destination['hour'] = pd.to_datetime(metar_destination.date, format='%d/%m/%Y %H:%M').dt.hour.astype('str')
metar_destination['date'] = pd.to_datetime(metar_destination.date, format='%d/%m/%Y %H:%M').dt.date.astype('str')

# sort by 'tail_number', 'date', 'hour' and drop duplicates
metar_destination = metar_destination.sort_values(['tail_number', 'date', 'hour']).drop_duplicates(subset=['tail_number', 'ICAO_destination', 'date', 'hour'])\
                           .reset_index(drop=True)

# inspect data
metar_destination.head()

# merge by 'tail_number', 'date', 'ICAO_origin', 'arr_UTC_hour'
rw = rw.merge(metar_destination.add_suffix('_destination'), how='left', left_on=['tail_number', 'date', 'ICAO_destination', 'arr_UTC_hour'],
        right_on=['tail_number_destination', 'date_destination', 'ICAO_destination_destination', 'hour_destination'])\
        .drop(['ICAO_destination', 'arr_UTC_hour','tail_number_destination', 'ICAO_destination_destination', 
               'date_destination', 'hour_destination'], axis=1)

# check for null values
rw.METAR_destination.isna().sum()

# percent of merged metar destination data
print(round(len(rw.loc[rw.METAR_destination.notnull()])/ len(rw) * 100,2), "%")

#### METAR Decoding
---

# decode metar data
rw = metar_decoder(rw)

# inspect data
rw.head()

rw.info()

rw.isna().sum()

# export as csv
rw.to_csv('../Datasets/NEW/RW_with_metar.csv', index=False)
