==> RW/Scripts/3 RW_anomaly.ipynb <==
#load packages
import pandas as pd 
import numpy as np
import matplotlib.pyplot as plt
pd.set_option('display.max_columns', None)

rw = pd.read_csv("../Datasets/NEW/RW_with_airports.csv", parse_dates=['date'])

rw.head()

rw.info()

rw.isna().sum()

# set Owner and tail_number as indexes
rw = rw.set_index(['Owner', 'tail_number'])

# calculate 25 quantile of each tail number of the airmedical service
rw['duration_q_25'] = rw.groupby(['Owner', 'tail_number']).duration.quantile(0.25)

# calculate 75 quantile of each tail number of the airmedical service
rw['duration_q_75'] = rw.groupby(['Owner', 'tail_number']).duration.quantile(0.75)

# calculate IQR
rw['duration_iqr'] = rw['duration_q_75'] - rw['duration_q_25']

# calculate lower whisker
rw['duration_lower_whisker'] = rw['duration_q_25'] - (1.5 * rw['duration_iqr'])

# calculate upper whisker
rw['duration_upper_whisker'] = rw['duration_q_75'] + (1.5 * rw['duration_iqr'])

rw['duration_outlier'] = np.where((rw.duration < rw.duration_lower_whisker) | (rw.duration > rw.duration_upper_whisker), "Yes", "No")

# calculate 25 quantile of each tail number of the airmedical service
rw['distance_q_25'] = rw.groupby(['Owner', 'tail_number']).distance_mi.quantile(0.25)

# calculate 75 quantile of each tail number of the airmedical service
rw['distance_q_75'] = rw.groupby(['Owner', 'tail_number']).distance_mi.quantile(0.75)

# calculate IQR
rw['distance_iqr'] = rw['distance_q_75'] - rw['distance_q_25']

# calculate lower whisker
rw['distance_lower_whisker'] = rw['distance_q_25'] - (1.5 * rw['distance_iqr'])

# calculate upper whisker
rw['distance_upper_whisker'] = rw['distance_q_75'] + (1.5 * rw['distance_iqr'])

rw['distance_outlier'] = np.where((rw.distance_mi < rw.distance_lower_whisker) | (rw.distance_mi > rw.distance_upper_whisker), "Yes", "No")

rw.head()

rw.to_csv('..//Datasets/NEW/RW_Outliers.csv')
