==> RW/Scripts/5. RW Tracker.ipynb <==
import pandas as pd
import numpy as np
pd.set_option('display.max_columns', 999)

# import data
rw = pd.read_csv("..//Datasets/NEW/RW_with_airports.csv")
rw_info = pd.read_csv("..//Datasets/NEW/RW_services_info.csv")

rw.head()

# number of existing tail numbers with flight history
rw.tail_number.nunique()

# The percentage of the total tail numbers
print(round(rw.tail_number.nunique() / rw_info.Tail_number.nunique() * 100, 2), "%")

rw.info()

rw = rw.iloc[::-1].reset_index(drop=True)

tail_number = []
date = []
time = []
airport = []
location = []
duration = []
distance_mi = []
Latitude = []
Longitude= []
Owner = []

for i in range(len(rw)):
    if rw.tail_number[i] == rw.tail_number.iloc[i-1]: 
        tail_number += [rw.tail_number[i]]
        date += [rw.date[i] + ' ' + str(rw.arrival[i])]
        airport += [rw.airport_origin[i]]
        location += [rw.city_origin[i]]
        Latitude += [rw.origin_Latitude[i]]
        Longitude += [rw.origin_Longitude[i]]
        Owner += [rw.Owner[i]]
        duration += [rw.duration[i-1]]
        distance_mi += [rw.distance_mi[i-1]]
    else:
        tail_number += [rw.tail_number[i]]
        date += [rw.date[i] + ' ' + str(rw.arrival[i])]
        airport += [rw.airport_origin[i]]
        location += [rw.city_origin[i]]
        Latitude += [rw.origin_Latitude[i]]
        Longitude += [rw.origin_Longitude[i]]
        Owner += [rw.Owner[i]]
        duration += [0]
        distance_mi += [0]
        
#     time += ['Departure']
    
#     tail_number += [rw.tail_number[i]]
#     date += [rw.date[i] + ' ' + str(rw.arrival[i])]
#     airport += [rw.airport_destination[i]]
#     location += [rw.city_destination[i]]
#     distance_mi += [rw.distance_mi[i]]
#     Latitude += [rw.destination_Latitude[i]]
#     Longitude += [rw.destination_Longitude[i]]
#     Owner += [rw.Owner[i]]
#     time += ['Arrival']
    
rw_tracker = pd.DataFrame(list(zip(tail_number, date, airport, location, duration, distance_mi, Latitude, Longitude, Owner)), 
             columns= ['tail_number', 'date', 'airport', 'location', 'duration', 'distance_mi', 'Latitude', 'Longitude', 'Owner'])    

# drop inconsequences
index = []
for i in range(len(rw_tracker)-1):
    if rw_tracker.airport.iloc[i] == rw_tracker.airport.iloc[i+1]:
        index += [i]
        
rw_tracker = rw_tracker.drop(index).reset_index(drop=True)

rw_tracker.to_csv('..//Datasets/NEW/RW_tracker.csv', index=False)

rw_tracker.head()

# create a dict
tail_number = {}

# add total flights and home_1 location
for i in rw_tracker.tail_number.unique():
    tail_number[i] = [len(rw_tracker.loc[rw_tracker.tail_number == i])]
    tail_number[i].append(rw_tracker.loc[rw_tracker.tail_number == i].airport.value_counts().index[0]) 

# create containers    
one_loc = 0
two_loc = 0
three_loc = 0
four_loc = 0

# loop over all paths 
for i in range(len(rw_tracker)):
    if rw_tracker.tail_number[i] == rw_tracker.tail_number.iloc[i-1]:
        if rw_tracker.airport[i] == rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[0]\
        and rw_tracker.airport.iloc[i-2] == rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[0]:
            one_loc += 1
        
        if rw_tracker.airport[i] == rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[0]\
        and rw_tracker.airport.iloc[i-3] == rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[0]:
            two_loc += 1
            
        if rw_tracker.airport[i] == rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[0]\
        and rw_tracker.airport.iloc[i-4] == rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[0]\
        and rw_tracker.airport.iloc[i-2] != rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[0]:
            three_loc += 1
            
        if rw_tracker.airport[i] == rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[0]\
        and rw_tracker.airport.iloc[i-5] == rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[0]\
        and rw_tracker.airport.iloc[i-2] != rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[0]\
        and rw_tracker.airport.iloc[i-3] != rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[0]:
            four_loc += 1
    else:
        tail_number[rw_tracker.tail_number.iloc[i-1]].append(one_loc)
        tail_number[rw_tracker.tail_number.iloc[i-1]].append(two_loc)
        tail_number[rw_tracker.tail_number.iloc[i-1]].append(three_loc)
        tail_number[rw_tracker.tail_number.iloc[i-1]].append(four_loc)
        
        # reset containers
        one_loc = 0
        two_loc = 0
        three_loc = 0
        four_loc = 0

# add home_2 location
for i in rw_tracker.tail_number.unique():
    try:
        tail_number[i].append(rw_tracker.loc[rw_tracker.tail_number == i].airport.value_counts().index[1])
    except:
        tail_number[i].append('None')

# create containers    
one_loc = 0
two_loc = 0
three_loc = 0
four_loc = 0

# loop over all paths 
for i in range(len(rw_tracker)):
    try:
        if rw_tracker.tail_number[i] == rw_tracker.tail_number.iloc[i-1]:
            if rw_tracker.airport[i] == rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[1]\
            and rw_tracker.airport.iloc[i-2] == rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[1]:
                one_loc += 1

            if rw_tracker.airport[i] == rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[1]\
            and rw_tracker.airport.iloc[i-3] == rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[1]:
                two_loc += 1

            if rw_tracker.airport[i] == rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[1]\
            and rw_tracker.airport.iloc[i-4] == rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[1]\
            and rw_tracker.airport.iloc[i-2] != rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[1]:
                three_loc += 1

            if rw_tracker.airport[i] == rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[1]\
            and rw_tracker.airport.iloc[i-5] == rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[1]\
            and rw_tracker.airport.iloc[i-2] != rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[1]\
            and rw_tracker.airport.iloc[i-3] != rw_tracker.loc[rw_tracker.tail_number == rw_tracker.tail_number[i]].airport.value_counts().index[1]:
                four_loc += 1
        else:
            tail_number[rw_tracker.tail_number.iloc[i-1]].append(one_loc)
            tail_number[rw_tracker.tail_number.iloc[i-1]].append(two_loc)
            tail_number[rw_tracker.tail_number.iloc[i-1]].append(three_loc)
            tail_number[rw_tracker.tail_number.iloc[i-1]].append(four_loc)

            # reset containers
            one_loc = 0
            two_loc = 0
            three_loc = 0
            four_loc = 0
    except:
        pass           

# create a pandas DataFrame
rw_hops = pd.DataFrame.from_dict(tail_number, orient='index',
                       columns=['total_flights','home_loc_1','one_loc', 'two_loc', 'three_loc', 'four_loc',
                               'home_loc_2','one_loc', 'two_loc', 'three_loc', 'four_loc'])\
            .reset_index().rename({'index': 'tail_number'}, axis=1)

# add city and state of location
rw_hops = rw_hops.merge(rw.drop_duplicates(subset=['airport_origin'])[['airport_origin', 'city_origin', 'origin_state']],
             how='left', left_on='home_loc_1', right_on='airport_origin').drop('airport_origin', axis=1)

rw_hops = rw_hops.merge(rw.drop_duplicates(subset=['airport_origin'])[['airport_origin', 'city_origin', 'origin_state']],
             how='left', left_on='home_loc_2', right_on='airport_origin').drop('airport_origin', axis=1)

# add owner
rw_hops = rw_hops.merge(rw_info[['Tail_number', 'Owner']], how='left',
             left_on='tail_number', right_on='Tail_number').drop('Tail_number', axis=1)

rw_hops

rw_hops.drop([894], axis=0).to_csv('..//Datasets/NEW/RW_hops.csv', index=False)
