==> RW/Scripts/6.1 RW_IHC.ipynb <==
import pandas as pd
import numpy as np
pd.set_option('display.max_columns', 999)

# load data frame
rw = pd.read_csv("..//Datasets/NEW/RW_with_airports.csv")

rw

# take only IHC service
ihc = rw.loc[rw.Owner == 'IHC HEALTH SERVICES INCSALT LAKE CITY, UT, US(Corporation)'].reset_index(drop=True)

ihc['day_name'] = pd.to_datetime(ihc.date, format='%Y-%m-%d').dt.day_name()

ihc['weekends'] = np.where((ihc.day_name == 'Saturday') | (ihc.day_name == 'Sunday'), 1, 0)

ihc.head()

ihc.to_csv('..//Datasets/NEW/RW_ihc.csv', index=False)

ihc.shape

# number of flights by tail number
ihc.tail_number.value_counts()

# number of flights by shift
ihc.departure_shift.value_counts()

# number of flights by shift
ihc.arrival_shift.value_counts()

# number of flights by shift
ihc[['distance_mi', 'duration']].mean()

# average distance and duration by tail number
round(ihc.groupby('tail_number')[['distance_mi', 'duration']].mean(),2)

# number of flights in origin and destination locations
ihc.groupby('tail_number')[['origin', 'destination']].nunique()

# top 5 of origin locations and number of flights for N281HC
ihc.groupby('tail_number').origin.value_counts().loc['N291HC'].head(5)

# top 5 of destination locations and number of flights for N281HC
ihc.groupby('tail_number').destination.value_counts().loc['N291HC'].head(5)

# top 5 of origin locations and number of flights for N481HC
ihc.groupby('tail_number').origin.value_counts().loc['N261HC'].head(5)

# top 5 of destination locations and number of flights for N481HC
ihc.groupby('tail_number').destination.value_counts().loc['N261HC'].head(5)

# distance summary statistics by departure shift
round(ihc.groupby('departure_shift').distance_mi.agg(['count','min','mean','median','max','std']),2)

# distance summary statistics by arrival shift
round(ihc.groupby('arrival_shift').distance_mi.agg(['count','min','mean','median','max','std']), 2)

# distance summary statistics by day
round(ihc.groupby('day_name').distance_mi.agg(['count','min','mean','median','max','std']), 2)

# duration summary statistics by departure shift
round(ihc.groupby('departure_shift').duration.agg(['count','min','mean','median','max','std']), 2)

# duration summary statistics by arrival shift
round(ihc.groupby('arrival_shift').duration.agg(['count','min','mean','median','max','std']), 2)

# duration summary statistics by day
round(ihc.groupby('day_name').duration.agg(['count','min','mean','median','max','std']), 2)

pd.DataFrame(ihc.origin.value_counts().head(5))

pd.DataFrame(ihc.destination.value_counts().head(5))

# Salt Lake City flights
slc = rw.loc[(rw.origin == 'Salt Lake City Intl (KSLC)') | (rw.destination == 'Salt Lake City Intl (KSLC)')\
            | (rw.origin == 'Salt Lake City, UT') | (rw.destination == 'Salt Lake City, UT')]

pd.DataFrame(slc.Owner.value_counts())

# Provo, UT flights
provo = rw.loc[(rw.origin == 'Provo, UT') | (rw.destination == 'Provo, UT')]

pd.DataFrame(provo.Owner.value_counts())

# competitors for N381HC in Salt Lake City as destination
pd.DataFrame(slc.loc[(slc.destination == 'Salt Lake City Intl (KSLC)') & (slc.tail_number != 'N481HC')].groupby('Owner')\
             .destination.value_counts().sort_values(ascending=False))

# competitors for N481HC in Salt Lake City as origin
pd.DataFrame(slc.loc[(slc.origin == 'Salt Lake City Intl (KSLC)') & (slc.tail_number != 'N381HC')].groupby('Owner')\
             .origin.value_counts().sort_values(ascending=False))

# competitors for N481HC in Salt Lake City as destination
pd.DataFrame(slc.loc[(slc.destination == 'Salt Lake City Intl (KSLC)') & (slc.tail_number != 'N381HC')].groupby('Owner')\
             .destination.value_counts().sort_values(ascending=False))

ihc.groupby('tail_number').destination.nunique()

# ihc.to_csv('ihc.csv', index=False)

# St George Rgnl (KSGU)
sgr = rw.loc[(rw.origin == 'St George Rgnl (KSGU)') | (rw.destination == 'St George Rgnl (KSGU)')]

# competitors for N381HC in St George Rgnl (KSGU) as origin
pd.DataFrame(sgr.loc[(sgr.origin == 'St George Rgnl (KSGU)') & (sgr.tail_number != 'N481HC')].groupby('Owner')\
             .origin.value_counts().sort_values(ascending=False))

# competitors for N381HC in St George Rgnl (KSGU) as destination
pd.DataFrame(sgr.loc[(sgr.destination == 'St George Rgnl (KSGU)') & (sgr.tail_number != 'N481HC')].groupby('Owner')\
             .destination.value_counts().sort_values(ascending=False))

# competitors for N481HC in St George Rgnl (KSGU) as origin
pd.DataFrame(sgr.loc[(sgr.origin == 'St George Rgnl (KSGU)') & (sgr.tail_number != 'N381HC')].groupby('Owner')\
             .origin.value_counts().sort_values(ascending=False))

# competitors for N481HC in St George Rgnl (KSGU) as destination
pd.DataFrame(sgr.loc[(sgr.destination == 'St George Rgnl (KSGU)') & (sgr.tail_number != 'N381HC')].groupby('Owner')\
             .destination.value_counts().sort_values(ascending=False))
